+++ src/Pages/index.ts
@@ -2,3 +2,4 @@ export { default as HomePage } from './HomePage';
 export { default as DetailPage } from './DetailPage';
 export { default as NotFoundPage } from './NotFoundPage';
 export { default as LoginPage } from './LoginPage';
+export { default as SignUpPage } from './SignUpPage';
